* More fixups for the ARM code analysis   - Fix a regression for certain non-load instructions   - Correctly resolve the deltas against 'pc' * XREFs are now displayed at from and dest addresses

+++ src/arch/arm/disarm.c
@@ -803,7 +803,7 @@ static CondPosT decode_unknown (Access32 instruction)
 
 
 /* Disassembler entry point */
-char *disarm (RawInstruction rawinstruction, int offset)
+char *disarm (RawInstruction rawinstruction, unsigned long long offset)
 {
 	CondPosT condpos;
 	Access32 instruction;
